@@ -1,7 +1,14 @@
 <?php
 
-if (PHP_VERSION_ID < 70000) {
-    date_default_timezone_set('Europe/Paris');
-}
+declare(strict_types=1);
+
+/*
+ * This file is part of the Arnapou jqCron package.
+ *
+ * (c) Arnaud Buathier <arnaud@arnapou.net>
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
 
 require __DIR__ . '/../vendor/autoload.php';
